refactor: Split connectivity.py into focused command modules
Split 1,521 LOC connectivity.py into 4 focused modules:

**New Modules:**
- connectivity_common.py (276 LOC) - Shared helpers
  - resolve_vm_identifier()
  - verify_vm_exists()
  - try_fetch_key_from_vault()
  - get_ssh_config_for_vm()

- connect.py (403 LOC) - SSH connection command
  - connect command with tmux/bastion support
  - Interactive VM selection
  - NFS quota checking

- file_transfer.py (615 LOC) - File transfer commands
  - cp command (multi-source file copy)
  - sync command (dotfile sync)
  - sync-keys command (SSH key sync)

- ide.py (288 LOC) - VS Code Remote launcher
  - code command with bastion tunnel support

**Improvements:**
- Each module < 600 LOC (meeting requirement)
- Single responsibility per module
- Extracted shared helpers to connectivity_common
- Backward compatibility via connectivity.py shim
- Deduped get_ssh_config_for_vm (was in both connectivity.py and env.py)

**Philosophy Compliance:**
- Bricks & Studs: Self-contained modules with clear contracts
- Zero-BS: All functions work, no stubs
- Ruthless Simplicity: One command per module

Resolves #1799
